// src/types.ts
export interface SearchResultProps {
	adult: boolean;
	backdrop_path: string | null;
	id: number;
	name: string;
	original_language: string;
	original_name: string;
	overview: string;
	title: string;
	poster_path: string | null;
	media_type: 'tv' | 'movie' | 'person';
	genre_ids: number[];
	popularity: number;
	first_air_date: string;
	vote_average: number;
	vote_count: number;
	origin_country: string[];
}
// ...
export interface suggestionProps {
	adult: boolean;
	backdrop_path: string;
	genre_ids: number[];
	id: number;
	original_language: string;
	original_title: string;
	overview: string;
	popularity: number;
	poster_path: string;
	release_date: string;
	title: string;
	video: boolean;
	vote_average: number;
	vote_count: number;
}

export interface suggestionsProps {
	date?: {
		maximum: string;
		minimum: string;
	};
	page: number;
	results: suggestionProps[];
	total_pages: number;
	total_results: number;
}